Extract split_recurring_series into impl
